requirements.txt
Django==5.1.6
gunicorn==23.0.0
psycopg2-binary==2.9.10
redis==4.5.5
celery==5.2.7
djangorestframework==3.15.2
requests==2.32.3
psutil==7.0.0
pillow
drf-yasg>=1.20.0
streamlink
python-vlc
yt-dlp
gevent==24.11.1
